[REFACTOR] Notification 엔티티 및 메세지 포맷터 이동
