+++ src/main/java/org/websoso/WSSServer/auth/validator/NotificationAuthorizationValidator.java
@@ -6,7 +6,7 @@ import static org.websoso.WSSServer.exception.error.CustomNotificationError.NOTI
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import org.websoso.WSSServer.domain.Notification;
+import org.websoso.WSSServer.notification.domain.Notification;
 import org.websoso.WSSServer.user.domain.User;
 import org.websoso.WSSServer.exception.exception.CustomNotificationException;
 import org.websoso.WSSServer.notification.repository.NotificationRepository;
