+++ src/main/java/org/websoso/WSSServer/feed/service/FeedService.java
@@ -30,8 +30,8 @@ import org.springframework.web.multipart.MultipartFile;
 import org.websoso.WSSServer.user.domain.AvatarProfile;
 import org.websoso.WSSServer.domain.Genre;
 import org.websoso.WSSServer.domain.GenrePreference;
-import org.websoso.WSSServer.domain.Notification;
-import org.websoso.WSSServer.domain.NotificationType;
+import org.websoso.WSSServer.notification.domain.Notification;
+import org.websoso.WSSServer.notification.domain.NotificationType;
 import org.websoso.WSSServer.user.repository.AvatarProfileRepository;
 import org.websoso.WSSServer.user.domain.User;
 import org.websoso.WSSServer.notification.domain.UserDevice;
