+++ src/main/java/org/websoso/WSSServer/feed/service/ReportService.java
@@ -30,7 +30,7 @@ import org.websoso.WSSServer.feed.repository.ReportedCommentRepository;
 import org.websoso.WSSServer.feed.repository.ReportedFeedRepository;
 import org.websoso.WSSServer.user.repository.UserRepository;
 import org.websoso.WSSServer.notification.service.DiscordMessageClient;
-import org.websoso.WSSServer.service.MessageFormatter;
+import org.websoso.WSSServer.notification.service.MessageFormatter;
 
 @Service
 @RequiredArgsConstructor
