+++ src/main/java/org/websoso/WSSServer/notification/repository/NotificationTypeRepository.java
@@ -3,7 +3,7 @@ package org.websoso.WSSServer.notification.repository;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import org.websoso.WSSServer.domain.NotificationType;
+import org.websoso.WSSServer.notification.domain.NotificationType;
 
 @Repository
 public interface NotificationTypeRepository extends JpaRepository<NotificationType, Long> {
