+++ src/main/java/org/websoso/WSSServer/notification/repository/ReadNotificationRepository.java
@@ -3,8 +3,8 @@ package org.websoso.WSSServer.notification.repository;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import org.websoso.WSSServer.domain.Notification;
-import org.websoso.WSSServer.domain.ReadNotification;
+import org.websoso.WSSServer.notification.domain.Notification;
+import org.websoso.WSSServer.notification.domain.ReadNotification;
 import org.websoso.WSSServer.user.domain.User;
 
 @Repository
