+++ src/main/java/org/websoso/WSSServer/notification/service/NotificationService.java
@@ -16,9 +16,9 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.websoso.WSSServer.domain.Notification;
-import org.websoso.WSSServer.domain.NotificationType;
-import org.websoso.WSSServer.domain.ReadNotification;
+import org.websoso.WSSServer.notification.domain.Notification;
+import org.websoso.WSSServer.notification.domain.NotificationType;
+import org.websoso.WSSServer.notification.domain.ReadNotification;
 import org.websoso.WSSServer.user.domain.User;
 import org.websoso.WSSServer.notification.domain.UserDevice;
 import org.websoso.WSSServer.domain.common.NotificationTypeGroup;
