+++ src/main/java/org/websoso/WSSServer/user/domain/User.java
@@ -17,13 +17,12 @@ import jakarta.persistence.UniqueConstraint;
 import java.time.Year;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.ColumnDefault;
 import org.websoso.WSSServer.domain.GenrePreference;
-import org.websoso.WSSServer.domain.ReadNotification;
+import org.websoso.WSSServer.notification.domain.ReadNotification;
 import org.websoso.WSSServer.feed.domain.ReportedComment;
 import org.websoso.WSSServer.feed.domain.ReportedFeed;
 import org.websoso.WSSServer.library.domain.UserNovel;
